PR Comments #2.0 (still needs testing)
