+++ cbxp/cbxp.h
@@ -7,8 +7,6 @@ extern "C" {
 #include <stdbool.h>
 #endif
 
-#include <stdint.h>
-
 typedef struct {
   char* result_json;
   int result_json_length;
